requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.1
attr==0.3.1
attrs==25.4.0
azure-core==1.38.2
azure-storage-blob==12.28.0
black==26.1.0
bleach==5.0.1
boto3==1.42.53
botocore==1.42.53
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.5
datamodel-code-generator==0.26.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.15
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==3.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-spectacular==0.28.0
email-validator==2.3.0
expiringdict==1.2.2
Faker==40.4.0
filelock==3.24.3
genson==1.3.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.5.0
google-cloud-logging==3.13.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.1
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsf==0.11.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
label-studio==1.22.0
label-studio-sdk==2.0.16
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.4.2
openai==1.109.1
opencv-python-headless==4.13.0.92
opentelemetry-api==1.39.1
ordered-set==4.0.2
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.3
psycopg-binary==3.3.3
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytokens==0.4.1
pytz==2022.7.1
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-file==3.0.1
requests-mock==1.12.1
rich==14.3.3
rpds-py==0.30.0
rq==2.6.1
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.16.0
semver==3.0.4
sentry-sdk==2.53.0
setuptools==82.0.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sqlparse==0.5.5
tldextract==5.3.1
toml==0.10.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==202602
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
user-agents==2.2.0
uuid_utils==0.14.0
webencodings==0.5.1
wheel==0.40.0
wrapt==2.1.1
xmljson==0.2.1
zipp==3.23.0
ultralytics==8.1.0
matplotlib
seaborn
jupyter
